requirements.txt
absl-py==0.14.0
astor==0.8.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
ckiptagger==0.2.1
clang==5.0
Cython==0.29.24
git+git://github.com/carpedm20/emoji@master#egg=emoji
filelock==3.0.12
flatbuffers==1.12
gast==0.2.2
gdown==3.13.1
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
idna==3.2
importlib-metadata==4.8.1
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
git+git://github.com/fossifer/markovify.git@make_sentence_that_contains#egg=markovify
mecab-python3==1.0.4
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
pkuseg==0.0.25
protobuf==3.18.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycld2==0.41
PySocks==1.7.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
Telethon==1.23.0
tensorboard==1.15.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==1.15.0
tensorflow-cpu==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
tqdm==4.62.3
typing-extensions==3.7.4.3
Unidecode==1.3.2
unidic-lite==1.0.8
urllib3==1.26.7
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0